validation

+++ server/app/src/chat/chat.service.ts
@@ -7,11 +7,43 @@ import {
 import { PrismaClient } from '@prisma/client';
 import { join } from 'path';
 import * as bcrypt from 'bcrypt';
+import { UserService } from 'src/user/user.service';
 
 const prisma = new PrismaClient();
 
 @Injectable()
 export class ChatService {
+  constructor(private userService: UserService) {}
+
+  async checkGroupPermissions(userId: number, chatId: number, admin: boolean) {
+    const chat = await prisma.chat.findFirst({
+      where: {
+        id: chatId
+      }
+    })
+
+    if (!chat)
+      return false
+
+    const userChat = await prisma.userChat.findFirst({
+      where: {
+        chatId: chatId,
+        userId: userId
+      },
+      select: {
+        role: true
+      }
+    })
+
+    if (!userChat)
+      return false
+    if (userChat.role == 'Admin' && !admin)
+      return false
+    if (userChat.role == 'User')
+      return false
+    return true
+  }
+
   async createGroup(
     userId: number,
     groupName: string,
@@ -42,6 +74,7 @@ export class ChatService {
   }
 
   async blockOrUnblock(userId: number, targetId: number, block: boolean) {
+    this.userService.getUserById(targetId)
     if (block) {
       await prisma.user.update({
         where: {
@@ -92,18 +125,8 @@ export class ChatService {
     targetId: number,
     chatId: number
   ) {
-    await prisma.userChat.update({
-      where: {
-        userId_chatId: {
-          userId: targetId,
-          chatId: chatId,
-        },
-      },
-      data: {
-        isMutted: true,
-      },
-    });
-    setTimeout(async () => {
+    this.userService.getUserById(targetId)
+    if (this.checkGroupPermissions(userId, chatId, true)) {
       await prisma.userChat.update({
         where: {
           userId_chatId: {
@@ -112,107 +135,108 @@ export class ChatService {
           },
         },
         data: {
-          isMutted: false,
+          isMutted: true,
         },
       });
-    }, 20 * 1000);
+      setTimeout(async () => {
+        await prisma.userChat.update({
+          where: {
+            userId_chatId: {
+              userId: targetId,
+              chatId: chatId,
+            },
+          },
+          data: {
+            isMutted: false,
+          },
+        });
+      }, 20 * 1000);
+    }
   }
 
   async kick(userId: number, targetId: number, chatId: number) {
+    this.userService.getUserById(targetId)
     if (targetId == userId)
       throw new BadRequestException('cant kick your self');
-    const chat = await prisma.userChat.delete({
-      where: {
-        userId_chatId: {
-          userId: targetId,
-          chatId: chatId,
+    if (this.checkGroupPermissions(userId, chatId, true)) {
+      const chat = await prisma.userChat.delete({
+        where: {
+          userId_chatId: {
+            userId: targetId,
+            chatId: chatId,
+          },
         },
-      },
-    });
+      });
+    }
   }
 
-  async ban(targetId: number, chatId: number) {
-    await prisma.userChat.update({
-      where: {
-        userId_chatId: {
-          userId: targetId,
-          chatId: chatId,
+  async ban(userId: number, targetId: number, chatId: number) {
+    this.userService.getUserById(targetId)
+    if (this.checkGroupPermissions(userId, chatId, true)) {
+      await prisma.userChat.update({
+        where: {
+          userId_chatId: {
+            userId: targetId,
+            chatId: chatId,
+          },
         },
-      },
-      data: {
-        isBanned: true,
-      },
-    });
+        data: {
+          isBanned: true,
+        },
+      });
+    }
   }
 
   async promote(userId: number, targetId: number, chatId: number) {
-    console.log(userId, '[]', typeof userId);
-    const userChat = await prisma.userChat.findFirst({
-      where: {
-        userId: userId,
-        chatId: chatId,
-      },
-    });
-    if (!userChat || userChat.role == 'User')
-      throw new UnauthorizedException(
-        'user have wrong permissions or not in group',
-      );
-
-    const targetChat = await prisma.userChat.findFirst({
-      where: {
-        userId: targetId,
-        chatId: chatId,
-      },
-    });
-
-    if (!targetChat) throw new UnauthorizedException('target not in group');
-
-    await prisma.userChat.update({
-      where: {
-        userId_chatId: {
+    this.userService.getUserById(targetId)
+    if (this.checkGroupPermissions(userId, chatId, false)) {
+      const targetChat = await prisma.userChat.findFirst({
+        where: {
           userId: targetId,
           chatId: chatId,
         },
-      },
-      data: {
-        role: 'Admin',
-      },
-    });
+      });
+
+      if (!targetChat) throw new UnauthorizedException('target not in group');
+
+      await prisma.userChat.update({
+        where: {
+          userId_chatId: {
+            userId: targetId,
+            chatId: chatId,
+          },
+        },
+        data: {
+          role: 'Admin',
+        },
+      });
+    }
   }
 
   async denote(userId: number, targetId: number, chatId: number) {
-    console.log(userId, '[]', typeof userId);
-    const userChat = await prisma.userChat.findFirst({
-      where: {
-        userId: userId,
-        chatId: chatId,
-      },
-    });
-    if (!userChat || userChat.role != 'Owner')
-      throw new UnauthorizedException(
-        'user have wrong permissions or not in group',
-      );
-
-    const targetChat = await prisma.userChat.findFirst({
-      where: {
-        userId: targetId,
-        chatId: chatId,
-      },
-    });
-
-    if (!targetChat) throw new UnauthorizedException('target not in group');
-
-    await prisma.userChat.update({
-      where: {
-        userId_chatId: {
+    this.userService.getUserById(targetId)
+    if (this.checkGroupPermissions(userId, chatId, false)) {
+      const targetChat = await prisma.userChat.findFirst({
+        where: {
           userId: targetId,
           chatId: chatId,
         },
-      },
-      data: {
-        role: 'User',
-      },
-    });
+      });
+
+      if (!targetChat) throw new UnauthorizedException('target not in group');
+
+      await prisma.userChat.update({
+        where: {
+          userId_chatId: {
+            userId: targetId,
+            chatId: chatId,
+          },
+        },
+        data: {
+          role: 'User',
+        },
+      });
+    }
   }
 
   async getUserByUsername(username: string) {
@@ -228,34 +252,22 @@ export class ChatService {
   async addToGroup(userId: number, target: string, groupId: number) {
     const user = await this.getUserByUsername(target);
     const targetId = user.userId;
-    const group = await prisma.chat.findFirst({
-      where: {
-        id: groupId,
-      },
-    });
-    const userChat = await prisma.userChat.findFirst({
-      where: {
-        userId: userId,
-        chatId: groupId,
-      },
-    });
-    const friendChat = await prisma.userChat.findFirst({
-      where: {
-        userId: targetId,
-        chatId: groupId,
-      },
-    });
-    if (friendChat) throw new BadRequestException('user already in group');
-    if (userChat.role != 'Owner' && userChat.role != 'Admin')
-      throw new UnauthorizedException(
-        'user is neither Owner or Admin in this group',
-      );
-    await prisma.userChat.create({
-      data: {
-        userId: targetId,
-        chatId: group.id,
-      },
-    });
+    
+    if (this.checkGroupPermissions(userId, groupId, true)) {
+      const friendChat = await prisma.userChat.findFirst({
+        where: {
+          userId: targetId,
+          chatId: groupId,
+        },
+      });
+      if (friendChat) throw new BadRequestException('user already in group');
+      await prisma.userChat.create({
+        data: {
+          userId: targetId,
+          chatId: groupId,
+        },
+      });
+    }
   }
 
   async leaveGroup(userId: number, groupId: number) {
@@ -271,15 +283,16 @@ export class ChatService {
         chatId: groupId,
       },
     });
-    if (!userChat) throw new BadRequestException('user not in group');
-    await prisma.userChat.delete({
-      where: {
-        userId_chatId: {
-          userId: userId,
-          chatId: groupId,
+    if (userChat) {
+      await prisma.userChat.delete({
+        where: {
+          userId_chatId: {
+            userId: userId,
+            chatId: groupId,
+          },
         },
-      },
-    });
+      });
+    }
   }
 
   async joinGroup(userId: number, groupId: number, password: string) {
@@ -288,11 +301,14 @@ export class ChatService {
         id: groupId,
       },
     });
+    if (!group)
+      return
     if (group.status == 'Private')
       throw new UnauthorizedException('group is private');
     if (group.status == 'Protected') {
+      if (!password || !password.length)
+        return
       const match = await bcrypt.compare(password, group.password);
-      console.log(password, '...', group.password, '...', match)
       if (!match)
           throw new UnauthorizedException('wrong password to joing group');
     }
@@ -304,12 +320,14 @@ export class ChatService {
     });
   }
 
-  async removeGroup(userId: string, groupId: number) {
-    const group = await prisma.chat.delete({
-      where: {
-        id: groupId,
-      },
-    });
+  async removeGroup(userId: number, groupId: number) {
+    if (this.checkGroupPermissions(userId, groupId, false)) {
+      await prisma.chat.delete({
+        where: {
+          id: groupId,
+        },
+      });
+    }
   }
 
   async changePass(userId: number, groupId: number, password: string) {
@@ -332,9 +350,9 @@ export class ChatService {
     if (
       !group ||
       group.status != 'Protected' ||
-      !password
+      !password || !password.length
     )
-      throw new InternalServerErrorException('something wrong');
+      throw new BadRequestException('something wrong');
     const salt = await bcrypt.genSalt();
     const hash = await bcrypt.hash(password, salt);
     await prisma.chat.update({
@@ -355,71 +373,67 @@ export class ChatService {
       },
     });
     if (!userChat) throw new UnauthorizedException('user not in group');
-    if (userChat.role != 'Owner' && userChat.role != 'Admin')
-      throw new UnauthorizedException(
-        'user is neither Owner or Admin in this group',
-      );
-    const group = await prisma.chat.findFirst({
-      where: {
-        id: groupId,
-      },
-    });
-    if (!group) throw new InternalServerErrorException('something wrong');
-    await prisma.chat.update({
-      where: {
-        id: groupId,
-      },
-      data: {
-        status: 'Public',
-        password: null,
-      },
-    });
+
+    if (this.checkGroupPermissions(userId, groupId, true)) {
+      await prisma.chat.update({
+        where: {
+          id: groupId,
+        },
+        data: {
+          status: 'Public',
+          password: null,
+        },
+      });
+    }
   }
 
   async createMessage(userId: number, chatId: number, content: string) {
-    const userChat = await prisma.userChat.findFirst({
-      where: {
-        userId: userId,
-        chatId: chatId
-      },
-      select: {
-        isMutted: true,
-        chat: {
-          select: {
-            isGroup: true,
-          },
+    if (content && content.length) {
+      const userChat = await prisma.userChat.findFirst({
+        where: {
+          userId: userId,
+          chatId: chatId
         },
-      },
-    });
-    
-    const chat = await prisma.userChat.findMany({
-      where: {
-        chatId: chatId,
-      },
-      select: {
-        isMutted: true,
-        chat: {
-          select: {
-            isGroup: true,
+        select: {
+          isMutted: true,
+          chat: {
+            select: {
+              isGroup: true,
+            },
           },
-        },
-        userId: true,
-      },
-    });
-
-    const isBlocked = !chat[0].chat.isGroup &&
-      ((await this.isBlocked(chat[0].userId, chat[1].userId)) ||
-        (await this.hasBlocked(chat[0].userId, chat[1].userId))) ||
-        userChat.isMutted;
-    console.log(isBlocked, 'is');
-    if (!isBlocked) {
-      await prisma.message.create({
-        data: {
-          senderId: userId,
-          chatId: chatId,
-          body: content,
         },
       });
+      
+      const chat = await prisma.userChat.findMany({
+        where: {
+          chatId: chatId,
+        },
+        select: {
+          isMutted: true,
+          chat: {
+            select: {
+              isGroup: true,
+            },
+          },
+          userId: true,
+        },
+      });
+      if (!userChat || !chat)
+        return
+      const isBlocked = !chat[0].chat.isGroup &&
+        ((await this.isBlocked(chat[0].userId, chat[1].userId)) ||
+          (await this.hasBlocked(chat[0].userId, chat[1].userId))) ||
+          userChat.isMutted;
+      // console.log(isBlocked, 'is');
+      if (!isBlocked) {
+        await prisma.message.create({
+          data: {
+            senderId: userId,
+            chatId: chatId,
+            body: content,
+          },
+        });
+      }
     }
   }
   async isBlocked(f1Id: number, f2Id: number) {
@@ -493,6 +507,8 @@ export class ChatService {
         createdAt: 'asc',
       },
     });
+    if (!messages)
+      return
     let fitlered = messages.filter(
       async (message) =>
         (await !this.isBlocked(userId, message.sender.profile.userId)) &&
@@ -504,7 +520,6 @@ export class ChatService {
       avatar: message.sender.profile.avatar,
       content: message.body,
     }));
-    // console.log( neededForm)
     return neededForm;
   }
 
@@ -563,6 +578,8 @@ export class ChatService {
         updatedAt: 'desc',
       },
     });
+    if (!chat)
+      return
     const neededForm = Promise.all( await chat.map(async (ch) => ({
       id: ch.chat.id,
       name: ch.chat.isGroup ? ch.chat.name : ch.dmName,
@@ -591,6 +608,7 @@ export class ChatService {
     if (!chat || chat.isBanned) return null;
     return chat.role;
   }
+
   async searchGroups(userId: number, keyword: string) {
     const matches = await prisma.chat.findMany({
       where: {
@@ -618,6 +636,8 @@ export class ChatService {
         status: true,
       },
     });
+    if (!matches)
+      return
     let filtered = matches.filter((m) => {
       const is = m.users.some((user) => user.userId === userId);
       return !is;
